# public/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarketMate API",
    "version": "2.1.0",
    "description": "The MarketMate API provides real-time and historical market intelligence for cryptocurrency and forex markets. Access market regime analysis, trading signals, liquidity maps, and historical data to power your trading applications, dashboards, and automated strategies.\n\n## Authentication\nAll API endpoints require a valid API key passed via the `Authorization` header as a Bearer token. You can generate and manage API keys from your MarketMate dashboard.\n\n## Rate Limits\n- **Free tier**: 60 requests/minute\n- **Pro tier**: 300 requests/minute\n- **Enterprise tier**: 1,000 requests/minute\n\nRate limit headers are included in every response: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`.\n\n## WebSocket Support\nReal-time data is also available via WebSocket at `wss://marketmate-website.onrender.com/ws`. See the event schemas documentation for WebSocket event formats.\n\n## SDKs & Tools\nOfficial SDKs are available for Python, JavaScript/TypeScript, and Go. Postman and Insomnia collections can be downloaded from the developers portal.",
    "contact": {
      "name": "MarketMate API Support",
      "url": "https://marketmate-website.onrender.com/developers",
      "email": "api@marketmate.io"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://marketmate-website.onrender.com/terms"
    }
  },
  "servers": [
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/market-state": {
      "get": {
        "operationId": "getMarketState",
        "summary": "Get current market state",
        "description": "Returns the current market regime analysis for a given symbol, including trend direction, conviction level, volatility assessment, and key support/resistance levels. Data is updated every 15 minutes.",
        "tags": ["Market Data"],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "description": "Trading pair symbol. Defaults to BTC-USD if not provided.",
            "schema": {
              "type": "string",
              "default": "BTC-USD",
              "enum": ["BTC-USD", "ETH-USD", "XAU-USD", "SOL-USD", "BNB-USD"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with market state data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketState"
                },
                "example": {
                  "symbol": "BTC-USD",
                  "regime": "trending_bullish",
                  "conviction": 0.82,
                  "volatility": "elevated",
                  "liquidity": 0.76,
                  "key_levels": {
                    "support": [67850, 67200, 66500],
                    "resistance": [69500, 70200, 71500]
                  },
                  "trend": {
                    "direction": "bullish",
                    "strength": 0.74,
                    "timeframe": "4H"
                  },
                  "updated_at": "2025-01-15T14:30:00.000Z"
                }
              }
            },
            "headers": {
              "X-RateLimit-Limit": {
                "description": "Maximum number of requests allowed in the current period",
                "schema": { "type": "integer" }
              },
              "X-RateLimit-Remaining": {
                "description": "Number of requests remaining in the current period",
                "schema": { "type": "integer" }
              },
              "X-RateLimit-Reset": {
                "description": "Time when the rate limit resets (Unix timestamp)",
                "schema": { "type": "integer" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "message": "API key required. Include Authorization: Bearer <your-api-key> header."
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                },
                "example": {
                  "error": "RateLimitExceeded",
                  "message": "You have exceeded your rate limit. Please retry after 60 seconds.",
                  "retry_after": 60
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                },
                "example": {
                  "error": "InternalServerError",
                  "message": "An unexpected error occurred. Please try again later."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/verify": {
      "get": {
        "operationId": "verifyApiKey",
        "summary": "Verify API key",
        "description": "Validates the provided API key and returns its status and environment. Use this endpoint to test your API key before making data requests. No rate limit is applied to this endpoint.",
        "tags": ["Authentication"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API key is valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                },
                "example": {
                  "valid": true,
                  "environment": "production"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyErrorResponse"
                },
                "examples": {
                  "missing_key": {
                    "summary": "No API key provided",
                    "value": {
                      "valid": false,
                      "error": "No API key provided"
                    }
                  },
                  "invalid_key": {
                    "summary": "Invalid or revoked key",
                    "value": {
                      "valid": false,
                      "error": "Invalid or revoked API key"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Pass your MarketMate API key as a Bearer token in the Authorization header."
      }
    },
    "schemas": {
      "MarketState": {
        "type": "object",
        "required": ["symbol", "regime", "conviction", "volatility", "liquidity", "key_levels", "trend", "updated_at"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Trading pair symbol",
            "example": "BTC-USD"
          },
          "regime": {
            "type": "string",
            "description": "Current market regime classification",
            "enum": ["trending_bullish", "trending_bearish", "ranging", "volatile", "transitional"],
            "example": "trending_bullish"
          },
          "conviction": {
            "type": "number",
            "description": "Conviction score for the current regime (0.0 - 1.0)",
            "minimum": 0,
            "maximum": 1,
            "example": 0.82
          },
          "volatility": {
            "type": "string",
            "description": "Current volatility assessment",
            "enum": ["low", "normal", "elevated", "extreme"],
            "example": "elevated"
          },
          "liquidity": {
            "type": "number",
            "description": "Overall liquidity score (0.0 - 1.0)",
            "minimum": 0,
            "maximum": 1,
            "example": 0.76
          },
          "key_levels": {
            "type": "object",
            "description": "Key support and resistance price levels",
            "required": ["support", "resistance"],
            "properties": {
              "support": {
                "type": "array",
                "description": "Support price levels (strongest first)",
                "items": {
                  "type": "number"
                },
                "example": [67850, 67200, 66500]
              },
              "resistance": {
                "type": "array",
                "description": "Resistance price levels (closest first)",
                "items": {
                  "type": "number"
                },
                "example": [69500, 70200, 71500]
              }
            }
          },
          "trend": {
            "type": "object",
            "description": "Current trend analysis",
            "required": ["direction", "strength", "timeframe"],
            "properties": {
              "direction": {
                "type": "string",
                "description": "Trend direction",
                "enum": ["bullish", "bearish", "neutral"],
                "example": "bullish"
              },
              "strength": {
                "type": "number",
                "description": "Trend strength score (0.0 - 1.0)",
                "minimum": 0,
                "maximum": 1,
                "example": 0.74
              },
              "timeframe": {
                "type": "string",
                "description": "Primary timeframe for the trend analysis",
                "enum": ["1m", "5m", "15m", "30m", "1H", "4H", "1D", "1W"],
                "example": "4H"
              }
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when this data was last updated",
            "example": "2025-01-15T14:30:00.000Z"
          }
        }
      },
...
      "VerifyResponse": {
        "type": "object",
        "required": ["valid", "environment"],
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Whether the API key is valid",
            "example": true
          },
          "environment": {
            "type": "string",
            "description": "The environment this API key is valid for",
            "enum": ["production", "sandbox"],
            "example": "production"
          }
        }
      },
      "VerifyErrorResponse": {
        "type": "object",
        "required": ["valid", "error"],
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Always false for error responses",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error description",
            "example": "Invalid or revoked API key"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type identifier",
            "example": "Unauthorized"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "API key required. Include Authorization: Bearer <your-api-key> header."
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "required": ["error", "message", "retry_after"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type identifier",
            "example": "RateLimitExceeded"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "You have exceeded your rate limit. Please retry after 60 seconds."
          },
          "retry_after": {
            "type": "integer",
            "description": "Seconds until the rate limit resets",
            "example": 60
          }
        }
      },
      "ServerError": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type identifier",
            "example": "InternalServerError"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "An unexpected error occurred. Please try again later."
          }
        }
      }
    }
  }
}